refactor: rewrite canvas helpers to ts
