+++ src/helpers/helpers.interpolation.ts
@@ -1,5 +1,4 @@
-import type {Point} from '../types/geometric.js';
-import type {SplinePoint} from './helpers.curve.js';
+import type {Point, SplinePoint} from '../types/geometric.js';
 
 /**
  * @private
