+++ src/helpers/types.ts
@@ -2,18 +2,4 @@
  * Temporary entry point of the types at the time of the transition.
  * After transition done need to remove it in favor of index.ts
  */
-
-// export * from '..js';
-export * from './helpers.color.js';
-export * from './helpers.collection.js';
-export * from './helpers.core.js';
-export * from './helpers.curve.js';
-export * from './helpers.dom.js';
-export * from './helpers.easing.js';
-export * from './helpers.extras.js';
-export * from './helpers.interpolation.js';
-export * from './helpers.intl.js';
-export * from './helpers.math.js';
-export * from './helpers.options.js';
-export * from './helpers.rtl.js';
-export * from '../types/helpers/index.js';
+export * from './index.js';
