--- src/types/helpers/index.d.ts
@@ -1,3 +1,0 @@
-export * from './helpers.canvas.js';
-export * from './helpers.canvas.js';
-export * from './helpers.segment.js';
